Rename status-constants to statuses
